feat: normalize and renaming of components
