+++ src/components/Layout/layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from '../Header/header';
+import Header from '../Header/Header';
 import ScrollToTop from '../../utils/ScrollToTop';
 import styles from './layout.module.css';
