+++ src/index.css
@@ -1,3 +1,5 @@
+@import url(./vendor/normalize.css);
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   font-size: 14px;
